switches span to contentEditable
since controlled inputs are meant to fire onChange events,
the old setup results in a ton of unnecessary transactions. Every
time the user changed a letter, it was updating on the dom. If switched to
onBlur, the transactions would fire but the dom would never update.

contendEditable comes with its own quirks, as its not recommended by React.
That said, it allows for update events to only fire onBlur while also updating
the dom.

* fixes spacing issue
* adds placeholder style for items